history-competition-ms/requirements.txt
flask
flask-cors
flask-swagger-ui
psycopg2-binary
redis
python-dotenv
